removed keyedmodel cache
